[Refactor] 타입관련 파일들 폴더에서 분리해, 파일로 관리
