+++ client/Hooks/Visitor/index.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "react-query";
 import { getVisitorAPI } from "services/Visitor";
-import { VisitorAPIType } from "Types/Visitor";
+import { VisitorAPIType } from "Types/visitor";
 import QUERY_KEY from "constants/queryKey";
 
 export const useGetVisitor = () => {
