--- client/Types/Dummy.ts
@@ -1,5 +1,0 @@
-export interface DummyType {
-	title: string;
-	createdAt: Date;
-	category: string;
-}
