+++ client/components/Atom/OtherPost/index.tsx
@@ -1,6 +1,6 @@
 import { useRouter } from "next/router";
 import React, { useCallback } from "react";
-import { OtherPostType } from "Types/Post";
+import { OtherPostType } from "Types/post";
 import styles from "./styles.module.scss";
 
 interface Props {
@@ -15,11 +15,7 @@ const OtherPost = ({ Post, mode }: Props) => {
     return push(`/post/${Post.OtherId}`);
   }, []);
   return (
-    <div
-      data-testid="OtherPost"
-      className={styles.OtherPost}
-      onClick={gotoPost}
-    >
+    <div data-testid="OtherPost" className={styles.OtherPost} onClick={gotoPost}>
       <div className={styles.description}>
         <span>{`${mode === "next" ? "다음" : "이전"}게시글`}</span>
       </div>
