+++ client/components/Block/Posts/index.tsx
@@ -1,6 +1,6 @@
 import { useRouter } from "next/router";
 import React from "react";
-import { PostsType } from "Types/Post";
+import { PostsType } from "Types/post";
 import PostCard from "components/Block/PostCard";
 import PostCardSkeleton from "components/Block/PostCard/Skeleton";
 import styles from "./styles.module.scss";
