+++ client/pages/index.tsx
@@ -5,7 +5,7 @@ import AdditionalInfoSectionRight from "components/Block/AdditionalInfoSectionRi
 import { useGetMainPost } from "Hooks/Post";
 import styles from "./styles.module.scss";
 import AdditionalInfoSectionLeft from "components/Block/AdditionalInfoSectionLeft";
-import { PostsType } from "Types/Post";
+import { PostsType } from "Types/post";
 import PostCard from "components/Block/PostCard";
 import { useInView } from "react-intersection-observer";
 import { useEffect } from "react";
